Editorial: decrease traversals for module graph fetching
Previously, we could invoke the internal module script graph fetching
procedure once for each combination of node and ancestor list, i.e. once
for each node and for each path to the node. This could be up to O(2^n)
for n nodes, as seen in
https://gist.github.com/domenic/d877f8b2e6d6ae62a9c94f916739e4db.

This change simplifies the model by using a single visited set per
top-level fetch, which ensures that the internal module script graph
fetching procedure is only invoked once for each node.

This change is technically editorial, as it has no visible, testable
effects. But it makes it much easier for implementers to follow the
specification as written without causing themselves massive slowdowns.